+++ packages/iov-bns/types/conditions.d.ts
@@ -1,6 +1,6 @@
 import { Address, ChainId, Hash, SwapId } from "@iov/bcp";
 import { As } from "type-tagger";
-/** A package-internal type representing a Weave Condition */
+/** A Weave condition */
 export declare type Condition = Uint8Array & As<"Condition">;
 export declare function buildCondition(extension: string, type: string, data: Uint8Array): Condition;
 export declare function buildMultisignatureCondition(id: number): Condition;